backend/requirements.txt
# ── Core Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# ── Database & Auth ──
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
cryptography>=42.0.0
bcrypt==4.2.0
python-multipart==0.0.18

# ── AI & LLM (Google Gemini) ──
google-generativeai==0.8.3

# ── Agents & Scraping ──
playwright==1.49.1
httpx==0.28.1
aiohttp>=3.9.0
aiofiles==24.1.0
apscheduler==3.10.4

# ── Redis ──
redis>=5.0.0

# ── Database Migrations ──
alembic>=1.13.0

# ── Data Processing ──
pypdf2==3.0.1
scikit-learn==1.6.0
numpy==2.2.0
jinja2==3.1.4

# ═══════════════════════════════════════════════════
# PHASE 11: Deploy & Production Dependencies
# ═══════════════════════════════════════════════════

# Task Queue (for async discovery/enrichment jobs)
celery==5.4.0
flower==2.0.0

# DNS / MX record validation (used by
# LeadEnrichmentAPI._guess_email_patterns for zero-cost
# email inference after Apollo/PDL waterfall miss).
dnspython>=2.4.0,<3.0.0

# Retry logic with exponential backoff (Phase 11 provider clients)
tenacity>=8.2.0,<9.0.0

# User-agent rotation (stealth scraping fallback path)
fake-useragent>=1.4.0,<2.0.0

# HTML parsing (already referenced by discovery_service)
beautifulsoup4>=4.12.0
